# Speedup property filtering