--- a/core/src/test/java/io/tracee/configuration/PropertiesBasedTraceeFilterConfigurationTest.java
+++ b/core/src/test/java/io/tracee/configuration/PropertiesBasedTraceeFilterConfigurationTest.java
@@ -1,5 +1,7 @@
 package io.tracee.configuration;
 
+import io.tracee.SimpleTraceeBackend;
+import io.tracee.TraceeBackend;
 import org.junit.Test;
 import org.mockito.Mockito;
 
@@ -18,24 +20,24 @@ import static org.mockito.Mockito.when;
 public class PropertiesBasedTraceeFilterConfigurationTest {
 
 	private PropertyChain propertyChain = Mockito.mock(PropertyChain.class);
-	private PropertiesBasedTraceeFilterConfiguration unit = new PropertiesBasedTraceeFilterConfiguration(propertyChain);
-
+	private TraceeBackend backend = SimpleTraceeBackend.createNonLoggingAllPermittingBackend();
+	private PropertiesBasedTraceeFilterConfiguration unit = new PropertiesBasedTraceeFilterConfiguration(backend.getLoggerFactory(), propertyChain);
 
 	@Test
 	public void testShouldPropagatePositive() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn(".*");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn(".*");
 		assertTrue(unit.shouldProcessParam("foo", AsyncDispatch));
 	}
 
 	@Test
 	public void testShouldPropagateNegative() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn("a,b,c");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn("a,b,c");
 		assertFalse(unit.shouldProcessParam("foo", AsyncDispatch));
 	}
 
 	@Test
 	public void testShouldAllowRegexInPatterns() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn("b[oa]+b");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn("b[oa]+b");
 		assertTrue(unit.shouldProcessParam("baab", AsyncDispatch));
 		assertTrue(unit.shouldProcessParam("boob", AsyncDispatch));
 	}
@@ -47,31 +49,31 @@ public class PropertiesBasedTraceeFilterConfigurationTest {
 
 	@Test
 	public void testShouldProcessIfDisabledByConfiguration() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn(" \t\n");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn(" \t\n");
 		assertFalse(unit.shouldProcessContext(AsyncDispatch));
 	}
 
 	@Test
 	public void testShouldProcessIfAnyPatternIsGiven() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn("a");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + AsyncDispatch.name())).thenReturn("a");
 		assertTrue(unit.shouldProcessContext(AsyncDispatch));
 	}
 
 	@Test
 	public void testGeneratedRequestIdLength() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + GENERATE_REQUEST_ID)).thenReturn("1");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + GENERATE_REQUEST_ID)).thenReturn("1");
 		assertThat(unit.generatedRequestIdLength(), equalTo(1));
 	}
 
 	@Test
 	public void testGeneratedRequestIdNonNumericMeansZero() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + GENERATE_REQUEST_ID)).thenReturn("false");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + GENERATE_REQUEST_ID)).thenReturn("false");
 		assertThat(unit.generatedRequestIdLength(), equalTo(0));
 	}
 
 	@Test
 	public void testGeneratedSessionIdLength() {
-		when(propertyChain.getProperty((TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + GENERATE_SESSION_ID))).thenReturn("42");
+		when(propertyChain.getProperty((TRACEE_DEFAULT_PROFILE_PREFIX + GENERATE_SESSION_ID))).thenReturn("42");
 		assertThat(unit.generatedSessionIdLength(), equalTo(42));
 	}
 
@@ -83,10 +85,8 @@ public class PropertiesBasedTraceeFilterConfigurationTest {
 
 	@Test
 	public void testFilterDeniedParamsPassesWhitelisted() {
-		when(propertyChain.getProperty(TRACEE_CONFIG_PREFIX + DEFAULT_PROFILE_PREFIX + IncomingRequest.name())).thenReturn("Foo");
+		when(propertyChain.getProperty(TRACEE_DEFAULT_PROFILE_PREFIX + IncomingRequest.name())).thenReturn("Foo");
 		final Map<String, String> unfiltered = Collections.singletonMap("Foo", "Bar");
 		assertThat(unit.filterDeniedParams(unfiltered,Channel.IncomingRequest), equalTo(unfiltered));
 	}
-
-
 }