Due caching of property filtering configuration we could skip string tokenizing and regex compilation and use a LRU cache instead.

With 2000000 filterings in a simple microbenchmark:
Pattern .* (permit all): old: 2seconds, new: 1,4seconds
Pattern tracee.* (permit keys prefixed with 'tracee'): Old: 5,8secs, new: 2,8secs